pyproject.toml
[project]
name = "ai-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "flask>=3.1.1",
    "ipython>=9.3.0",
    "langchain[google-genai]>=0.3.25",
    "langgraph>=0.4.8",
    "langsmith>=0.3.45",
    "pip>=25.1.1",
    "python-dotenv>=1.1.0",
    "pywebview>=5.0.0",
    "PyQt5>=5.15.0",
    "PyQtWebEngine>=5.15.0",
    "QtPy>=2.0.0",
]
